Make build_tools not to be built when all is built
There are build targets 'all' and 'build_tools'. When 'all' is built,
build 'build_tools' is required. But 'build_tools' makes an binary file
called bin2h. This is an architecture dependent object. This is useless.
And lintian check is failed, because of it. So make 'build_tools' disabled
when 'all' is built.