Remove modules.order modification code
Previously, the module was placed at first line in modules.order.
I did it so that the kernel loaded the module the first order among another
modules what have aliases for rt3070.
But what I did is not perfect solution to make the kernel to load the module
the first order. It does not work properly when the kernel is updated.

This commit reverts c82931e, eab7dfb and 8095a9c.